Use functions instead of closures
Grammars can now recur!
The switch from closures to functions is made by introducing a new macro syntax `nonterm: type => ...`.
The type is used to determine the return type of the function.
The functions are *not* held in a `HashMap` like earlier versions of the parser.
A drawback is that functions can't capture scopes but this should not be an often desired behavior.

Add a new test for recursive grammars.